Add files via upload
1 Shortest Common String Time complexity 
The specific part of the brute force solution that gives it an undesirable time 
complexity is the use of permutations. Generating all possible permutations of a 
set of n strings has a time complexity of O(n!). This means that the time taken to 
generate all permutations increases extremely rapidly as the size of the input set 
grows. 
Furthermore, the overlap function has a time complexity of O(m2). where m is the 
length of the strings being compared. The overlap function is called within the 
shortestCommonSuperstring function for each pair of adjacent strings in the 
permutation, resulting in a total time complexity of O(n2 ∗m2). This is also 
computationally expensive and can result in long runtimes for large inputs. 
Therefore, the brute force solution has an undesirable time complexity of O(n! * 
n2∗m2), making it impractical for large inputs. 
The time complexity of your improved solution implemented in the 
”scsfast.py” file is. O(n2k2), 
2 De Brujin Graphs 
2.1 Generating Your Own Unique Data 
Generated unique data set from the provided code datasetGenerator hw2.py. Ran 
the above program and provide student ID (1002068038) as one of the input. 
Then run the command line python3 datasetGenerator hw2.py –ID 1002068038. 
A new Text document will be created as 1002068038.txt with unique reads in it. 
2.2 Generating K-Mers 
Calculated K-mers with k=2 and generated the strings in new file named kmers.txt 
1 
2.3 Generating a De Bruijn Graph 
Taking the available 1002068038.txt as input, a De Brujin graph is generated with 
the available string in the text file with k=2. 
2.4 Eulerian Cycle 
Since the De Bruijn graph doesn’t follow the Eulerian walk i.e it doesn’t cover all 
the edges, my De Bruijn graph is not an Eulerian cycle. 
2.5 Genome Assembly 
Since the graph is not an Eulerian cycle, There is a need to add new reads to the 
graph, the minimum set of reads required for my De Bruijn graph to convert to 
Eulerian cycle is 1(new read will be added as an edge between ’gt’ and ’tc’ which 
is previously not present), and one of the assembled genome sequence formed 
from the above reads will be gacgcgtcgag 
2.6 Difficulty Adjustment 
It took 15-16 hours for me to complete the assignment. 
writing the best code for the shortest common superstring and finding its time 
complexity took time and Finding assembled genome sequence took some more 
time. 
2