Add files via upload
1 Substitution Matrix 
Given that, A ←→G and T ←→C) are less common than tranversions (A ←→ 
T , A ←→ C, G ←→ T , and G ←→ C) 
The substitution matrix for above changes is 
[[X A G T C] 
[A 1 -2 -1 -1] [G -2 1 -1 -1] 
[T -1 -1 1 -2] 
[C -1 -1 -2 1]] 
2 Global Alignment 
Using Needleman-Wunsch algorithm I constructed a 2D matrix D with the help of 
function named global alignment takes variables,Sequence A,Sequence 
B,substitution matrix,gap penalty 
Example: Sequence A=”GATA” Sequence=”CTAC” with a match score of 1, 
mismatch score of -1, and gap score of -2. 
The D matrix is generated as below: 
[[ 0, -2,-4, -6, -8] 
[-2, -1, -3, -5, -7] 
[-4, -3, -2, -2, -4] 
[-6, -5, -2, -3, -3] 
[-8, -7, -4, -1, -3,]] 
As we traversing D,for the input sequence “GATA”and “CTAC” we get the global 
alignments such as [(”GATA-”,”C-TAC”),(”GATA-”,”-CTAC”)] 
Example: Sequence A=”AATCA” Sequence B=”AACG” with a match score of 1, 
mismatch score of -1, and gap score of -2. 
The D matrix is generated as below: 
[[0, -2, -4, -6, -8] 
[-2, 1, -1, -3, -5] 
[-4, -1, 2, 0, -2] 
[-6, -3, 0, 1, -1] 
1 
[-8, -5, -2, 1, 0] 
[-10, -7, -4, -1, 0]] 
we get the global alignments such as [(”AATCA”, ”AA-CG”)] 
3 Local Alignment 
Using Smith-Waterman algorithm constructed a 2D matrix D with the help of function 
named local alignment takes variables,Sequence A,Sequence B,substitution 
matrix,gap penalty 
Example: Sequence A=”CGATA” Sequence B=”CAGTA” with a match score of 1, 
mismatch score of -1, and gap score of -2. 
The D matrix is generated as below: 
[[0, 0, 0, 0, 0, 0] 
[0, 1, 0, 0, 0, 0] 
[0, 0, 0, 1, 0, 0] 
[0, 0, 1, 0, 0, 1] 
[0, 0, 0, 0, 1, 0] 
[0, 0, 1, 0, 0, 2]] 
As we traversing D,for the input sequence “CGATA” and “CAGATA” we get the 
global alignments such as [(”TA”, ”TA”)] 
Example: Sequence A=”GTACA” Sequence B=”ATCAA” with a match score of 1, 
mismatch score of -1, and gap score of -2. 
The D matrix is generated as below: 
[[0, 0, 0, 0, 0, 0] 
[0, 0, 0, 0, 0, 0] 
[0, 0, 1, 0, 0, 0] 
[0, 1, 0, 0, 1, 1] 
[0, 0, 0, 1, 0, 0] 
[0, 1, 0, 0, 2, 1]] 
As we traversing D,for the input sequence ”GTACA” and ”ATCAA” we get the global 
alignments such as [(”CA”, ”CA”)] 
4 Custom Alignment 
Taking first name and last name in lowercase and concatenating them, 
firstname=”revanth” ,lastname=”kurmeti” when we concatenate string is 
”revanthkurmeti” , 
2 
With the help of function named local alignment from above problem which takes 
variables,Sequence A,Sequence B,substitution matrix,gap penalty match=2,semi
match=+1,mismatch=-1,gap=-2 Sequence A=”revanthkurmeti” 
Sequence B=”thequickbrownfoxjumpsoverthelazydog” 
The score matrix is included in file 1002068038 S.txt 
The pretty print is included in 1002068038 D.txt 
The local alignment for the above strings is [(’vanthk’, ’verthe’), (’vanthkur’, 
’verthela’)] 
3